NTBS-731: Merge immunosuppression and comorbities edit pages
