+++ ntbs-service/Pages/Shared/_SideMenu.cshtml
@@ -48,11 +48,6 @@
                 Co-morbidities
             </a>
         </li>
-        <li class="@("contents-list-item " + (currentPage == NotificationSubPaths.EditImmunosuppression ? "contents-list-item--current" : ""))">
-            <a class="app-subnav__link govuk-link govuk-link--no-visited-state" href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.EditImmunosuppression)">
-                Immunosuppression
-            </a>
-        </li>
         <li class="@("contents-list-item " + (currentPage == NotificationSubPaths.EditSocialContextAddresses ? "contents-list-item--current" : ""))">
             <a class="app-subnav__link govuk-link govuk-link--no-visited-state" href="@RouteHelper.GetNotificationPath(Model.NotificationId, NotificationSubPaths.EditSocialContextAddresses)">
                 Social Context - Addresses
